+++ apps/probability/law/exponential_law.h
@@ -18,9 +18,9 @@ public:
   I18n::Message parameterNameAtIndex(int index) override;
   I18n::Message parameterDefinitionAtIndex(int index) override;
   float evaluateAtAbscissa(float x) const override;
-  bool authorizedValueAtIndex(float x, int index) const override;
-  float cumulativeDistributiveFunctionAtAbscissa(float x) const override;
-  float cumulativeDistributiveInverseForProbability(float * probability) override;
+  bool authorizedValueAtIndex(double x, int index) const override;
+  double cumulativeDistributiveFunctionAtAbscissa(double x) const override;
+  double cumulativeDistributiveInverseForProbability(double * probability) override;
 };
 
 }
